Refactored getlistings_uk_ananova's channel management into Ananova_Channel.pm.  Previously there were too many hashes mapping between XMLTV channel id, Ananova channel id, Ananova display name (which comes in two forms), additional display name, broadcast method or region, channel packages, and what the user wishes to download. It's a little bit saner now but still rather ugly.
The Ananova_Channel package is not in any way a general or reusable
set of channel management routines.  It's very specific to the
peculiarities of this data source.  The name sucks, but we're going to
rename everything soon anyway.

Also a few bug fixes including that the episode-num is now written out
properly (at least until that bit of DTD gets replaced with something
saner), updates to channel definitions and packages, and general
tidying.

The old option ALL to get every one of Ananova's pages will probably
become undocumented in the next release.  It doesn't work that well
because often two different pages on Ananova are mapped to the same
XMLTV channel.  It ends up grabbing both.  This is however useful for
checking the output, running it through sort_listings to check that
the two Ananova channels are indeed identical, so I won't remove this
ability completely.