Begun the process of moving the hardcoded channel data out of tv_grab_uk and into text files.  To start with I have put the big list mapping Ananova ids to XMLTV ids into its own file, which lives in the share/ directory (eg /usr/local/share/).  This meant altering Makefile.PL to support 'share' files in the same way it was already altered for 'docs', and processing tv_grab_uk on installation to set the path to the data file.
One side effect of the new way of loading the data is that the channel
'BBC1' now maps to generic BBC1 rather than the BBC1 West it was before
(for some reason).  This is probably because I reordered the Ananova ids
in the channel_ids file.  But that looks like a good side effect.