Well, up till now I've resisted calls to trim the number of modules required by xmltv.  I said the answer was not to reduce library use but rather to make it easy for users to install those libraries, by providing them as packages.  But after moving to a new system myself I realize that this attitude was too optimistic.  Whatever the merits in principle of packaging the libraries separately, in practice it is a lot of effort to keep track of those dependencies.
As a starting point I have reduced the dependencies to packages which
come as standard with Linux-Mandrake 8.2.  This meant removing
dependencies on obscure modules (many of which I wrote myself).  Some,
like Tie::IxHash and Sort::Versions, turned out not be used anywhere,
although they were once.  Others like Log::TraceMessages and Memoize
can be replaced with stub routines when they are not found, and there
are some like Lingua::Preferred where doing this gives a small loss of
functionality, but not anything that most users will notice.

On the other hand, I noticed that tv_pick.cgi requires the CGI module,
which had slipped through the dependency net until now :-(.  At least
it is a common module which will probably be packaged (it is in
Mandrake).

I've changed the Makefile.PL and README accordingly; building will
warn about 'recommended' modules which are missing but they don't stop
the Makefile being written.

I've run many of the altered programs on my own machine but more
testing is needed.

Also improved the 'please run --configure' message in tv_grab_uk.