Made sure the XML is valid by cleaning up ('demoronizing') a couple of non-Latin-1 characters which were being carried through from the HTML pages.
Removed retry-fetching, I may add it back later if it proves to be
needed.  But it interferes with caching which is a nuisance for
development.

Changed non-standard keys in hashes to begin with an underscore, so
XMLTV::Writer will silently skip them rather than complaining.

Added some more error checking, in particular a check that regular
expression matches actually succeed before we go off and use captured
strings.

Strip leading and trailing whitespace from titles and descriptions.

Rejigged the channel handling a little; now we try to use XMLTV ids
internally and only convert them to Radio Times ids when contacting
the site.  Also, only the channels which are actually used are written
out as <channel> elements, and with the correct id.  Still some more
to do here, in particular a real configuration stage.

Added prototypes for all subroutines.

Cosmetic formatting, indentation and code style changes.

Removed use of Time::ParseDate, now we use Date::Manip exclusively for
that stuff.