The name regular expressions were made weaker to accommodate names like 'LL Cool J' (sigh).  But I added a negative lookbehind test to make sure that 'the' does not precede a name unless it is part of the name.  How many of these lookbehind and lookahead tests should be part of $NAME_RE, and how many part of the regexps which embed it, is still to be sorted out.
special_category(): report more accurately whether the programme has
been changed.

special_credits(): work on all the descriptions in the right language,
not just the first.  Improved some of the extractor regexps, and more
extractors now examine categories, or the presence of an existing
person with the same job, to see whether they should run.  Did away
with the silly (non?)optimization of 'caches' stored in the programme
details listing people in a given job.  It is not possible for a
person to both be a guest and do some other job.