Added function prototypes.
get_courteous(): there seems to be a difference in Perl versions on
how calling get_courteous_aux(@_) is interpreted.  Or some weird
interaction with Memoize.  It seems that @_ was being turned into a
scalar by taking the length of the list, ie 1.  I've changed it to
call get_courteous_aux($_[0]), which works on perl 5.8 and should work
on all other versions.