Pretty much a total rewrite.
On my development box (and hence the resulting EXEs) application DIE messages
were not being delivered to the users.

This version is much simpler (no self generating code), uses a "require"
instead of a "do" and seems to work fine.  Error statuses are even passed
as DOS ERRORLEVELS if you case to script it.