Fixed a design bug in tv_sort.  Previously, when comparing two programmes, if either of them lacked a stop time then it did not consider stop times and went straight on to comparing other things. But this makes a bad comparison function: for example three programmes which have the same start time, one of them has no stop time, and the other two have different stop times.  You can see that the <= relation is no longer transitive.  Fixed this by deciding that 'no stop time' comes before 'some stop time'.
This meant some changes to add_stop(), the routine which guesses stop
times for programmes that don't have them.  This function ought
perhaps to be in a separate filter program, but it's still part of
tv_sort for now.  add_stop() needs to maintain the invariant that its
list of programmes is sorted, so you can't always add stop times
because that could now change the sorting order.  But I've rewritten
it so it's a bit less tangled than before, I'm fairly sure it's
correct, and the results seem about the same as before (except in
cases where the old version broke).