Fixed manual page generation (again).  For some reason $(INSTALLMAN1DIR), which is normally the final resting place of section 1 manual pages, wasn't created at the time windows_dist ran. Instead we now fetch the pages from $(INST_MAN1DIR), which is the temporary place in blib/, and should always be there.  I hope this doesn't break on anyone else's system with a different MakeMaker.
The 'unix2dos' program also has a tendency to misbehave - it can run
in two modes, modifying a file in-place or running as a filter and
writing the new text to stdout.  My unix2dos seems to have decided to
switch from one to the other, so I've replaced it with some Perl code
to do the same job.