Added <tz-hint> which lets applications convert times into an appropriate timezone to show to the user, even though the times in the XML file are always UTC.  (But documented that code-times for systems like VPS and PDC, which are really just magic numbers, may be in a local timezone.)
Decided that newline is treated the same as space; this is consistent
with HTML and a bit saner than the old DTD's prohibition of newline.

Added an 'id' element to <programme> and created <programme-ref> to
refer to it.  Programme-refs can appear inside a timeslot the same way
programmes can.  But they must refer to a programme defined earlier in
the file (to allow for lazy reading).

Stopped misuse of <desc> inside <link> - it was a good idea at first,
but then we added 'type of desc' which made it specific to programmes.

Documented the <number> element, it was there all along but I hadn't
decided what it meant.  It can store any kind of channel number,
usually the button you'd press on your television, but also
application-defined things depending on the 'system' attribute.  I
don't really want to make it any more complicated than that.