get_channels(): read the channel names file.  The new boolean parameter says whether to read the site as well and reconcile the two (more paranoia, but also more downloading).
Call get_channels() to get channel details before grabbing as well as
for --configure and --list-channels.  --configure still downloads the
channel list from the site but --list-channels and grabbing do not.
Changed the error about bad channel names to a warning, since it can
happen if the channels file has different names from the site.