Fixed for site changes.  Previously I think the site wrote programme times as images, but now they're just bold text.  (The output was giving every programme time as January 1st 1970.)
Added --offset option.  It was included in the usage message but not
accepted.  However, because the site doesn't allow grabbing with an
offset, the option just extends the range of days grabbed.  So if you
ask for 'two days starting tomorrow' you will in fact get the next
week, since that's the shortest range including what you requested.
Also reworked the number of days code a little bit in general, so that
a --days flag on the command line overrides any config file setting.

The grabber has its own simple --cache mechanism but this was never
used since XMLTV::Memoize::check_argv() gobbles any --cache flag given
on the command line.  And the function memoized wasn't ever called.  I
have removed the check_argv() call so now caching does something;
however it needs to be revisited because writing known filenames in
/tmp/ is a security hole.

Added a little more error checking, and try to include the URI in
some error messages.

New dependency: List::Util.  I'll check on the mailing list about
this.