tv_grab_fr: Add (initial) support for new telestar.fr listings site
This is a significant update and requires channels to be reconfigured
via tv_grab_fr --configure. Sorry.

The new telestar.fr site supports 160+ channels with 14 days of data.

Most of the available programme data is parsed and made available in the
XML output. I'm sure there are data which could be included that the script
does not currently handle. Details/patches appreciated.

Please report any issues to the xmltv-users mailing list.