tv_grab_sd_json: Improve efficiency of date/time handling.
After profiling the code using NYTProf I found that the DateTime parsing
and comparisons were the top usage of execution time. Adding my own parsing
functions and converting everything to seconds from epoch for comparisons
makes a huge difference in performance. Excluding network delays (or when
everything is already in the cache) the execution time is now about cut in
half.

DateTime usage is actually still the number one item in the profiler output
so there may still be room for improvement here but dates and times are
hard to get right with all the timezone, DST, leap, etc. issues.