Do not hardcode usage of JSON::XS when depending on JSON
tv_grab_huro declares a runtime dependency on JSON, but hardcodes calls to
JSON::XS in the actual code. JSON::XS is not declared as a build-time dependency
so there is not guarantee is will be installed.

Therefore, we declare a build-time dependency on JSON and change calls to use
the JSON wrapper, which will use either JSON::PP or JSON::XS if available.