Ensure tzdata package is installed on recent Ubuntu images
Recent Ubuntu docker images removed the tzdata package. Date::Manip
version 6.57 (and possibly some earlier versions) contain a bug that
manifests when /etc/timezone is missing (or is set to an invalid timezone).
This is possible when tzdata is not installed.

Date::Manip version 6.57 is included in the Ubuntu 17.04 "Zesty Zapus"
distro which *does not* include tzdata by default.

Date::Manip version 6.57 is included in the Debian 9 "Stretch" distro
which *does* include tzdata by default.

The relevant bug was fixed in Date::Manip 6.58+. See:
- https://github.com/SBECK-github/Date-Manip/issues/8
- https://github.com/SBECK-github/Date-Manip/issues/9
for more details.