Migrate remaining scripts to "use JSON;" wrapper by default
JSON is a wrapper for JSON::XS-compatible modules. Several grabbers
already specify its use by default. The remaining grabbers and libraries
with a requirement for processing JSON data hardcode the use of either
JSON::PP or JSON::XS.

JSON uses JSON::XS by default, and when JSON::XS is not available, falls
back to JSON::PP, which is in the Perl core since 5.14. If JSON::PP is
not available (systems running Perl <5.14), JSON falls back on
JSON::backportPP which is bundled with JSON.  It is also possible to
explicitly specify the use of Cpanel::JSON::XS if desired (a fork of
JSON::XS by Reini Urban).

For those grabbers not already using the JSON wrapper, update their
scripts and their declared dependencies to use JSON instead.

Add details to the README and promote JSON to a core dependency.