Download files to sane temporary files
Just use fixed file-names instead of using the name from
Content-Disposition header and sanitizing it.

The real name is only needed for extracting the channel group and for
nothing else. (I find the name only useful for debugging and probably
for caching, but this doesn't work with epgdata.com)

Also download directly to a file instead of going through the RAM.