tv_grab_na_dtv: add timeout/agent string options
This commit is an amended version of ejones's Pull Request #82, which
addresses #80 following reports of the grabber ceasing to retrieve data.

This update adds local timeout and useragent string options to the grabber
to allow overriding the default LWP/XMLTV values of 180s and
"xmltv/$XMLTV::VERSION", respectively.

The default timeout for the grabber is raised to 240 seconds.

Closes: #80
Closes: #82

Signed-off-by: Nick Morrott <knowledgejunkie@gmail.com>