XMLTV.pm: Read from STDIN using XML::Twig->parse()
Recent changes to XML::Parser to use 3-arg open removed its support for
reading from STDIN when the special filename of '-' was passed.

Update XMLTV::parsefiles_callback() to check for the presence
of the special '-' filename from calling applications and call
XML::TWig->parse() or XML::Twig->parsefile() as appropriate.

Closes: #77
Closes: #78