Reorganize database open/validation
Move some of the database creation steps inside the
version upgrade process (version 0 to 1), change the
initial test to not require a write lock, and only
try to change database characteristics during initial
creation (when they are typically allowed to change).
This also supports someone to manually configure
certain characteristics and have them left as set.