increase potential grabber concurrency phase 1
Eliminate some unnecessary write locks, and
try not to have the database locked across
Schedules Direct requests (some are quick,
some may take longer).

While sqlite appears to preserve statement
handles across a commit not all databases
do and it is not assured that sqlite will
continue to do so, so refactor the statement
prepares to be within loops that now have
commits.