Correct tv_grab_zz_sdjson to be XMLTV DTD compliant
Per discussions on the xmltv-devel list (starting
on 2020-04-13, subject "Request for clarification"),
it was agreed by the project elders that new episodes
of programs cannot have a previously shown, and
that new is only for true first showings of the
first episode (the Schedules Direct definition of
new, which is every new episode of every series
is not the same as the XMLTV definition of new).

The tv_grab_zz_sdjson grabber is not XMLTV compliant.

Note that the (deleted in this patch) comment about
MythTV requiring the original airdate is no longer
applicable in the latest MythTV release (i.e. it
adjusts the data per XMLTV definitions as necessary).

Closes: #91