support ordering of station logos
Schedules Direct recently reorganized their
station logos.  There are now (generally)
four variants of a station logo (there used
to be primarily two), which are categorized
as "dark" for use on a dark background (fixed
text and such tends to be white), "light" for
use on a light background (fixed text tends
to be black), "gray" which is grayscale and
typically used with a light background, and
"white" which is white and typically used
on dark backgrounds.

The default order from Schedules Direct will
currently result in the order being dark,
gray, light, white when placed in the XMLTV
channel output.

Some applications will only process/store
the first returned logo, so dark it often is.

We now allow the user to run --configure to
change the order of the returned data to be
as desired.  The actual internal code allows
one to reorder the list in a specific order,
but --configure only allows one to specify a
single priority category (--configure is not
designed to support complex ordering
selection criteria), which should be adequate
for apps that only process the first logo
(those that process all logos and allow user
selection likely will chose to specify no
additional ordering).

The full syntax of the configuration file
option is:

  station-logo-order=Source/category[,Source/category]*

Right now the only source is Gracenote with
the categories dark, gray, light, and white.