build.yml: add Alpine Linux 3.17
Alpine Linux has a community package for XMLTV. So it is a good idea to
add it to our build flow too.

Unfortunately not all Perl dependencies are available as ready-made
packages, so we have to build some straight from CPAN. Currently missing
is SOAP::Lite, which fails to build.