yle: rewrite grab()
The data for this source is only available via an API that requires
authentication. But we do not want to include secrets in the source
code.

If a user wants use this source, then he needs to fire up the browser,
open the developer tools window, go to the web site and figure out the
API key from the network traces generated by the web site.

Add yle.fi options to test configuration to guide the user what he needs
to look up with the browser.