Support github actions CI on legacy distros requiring Node 16
Github has noted that Node 16 has reached end of life, and
has announced deprecation of Node 16 support for github
actions by Spring of 2024 (with a transition to Node 20).

The project still has a few distros in the CI chain that cannot
support Node 20 (and never will, as they are either beyond,
or will shortly be, EOL).

As an interim step, use the legacy checkout actions that
only require Node 16 for those distros.  Note that using
Node 16 will result in a warning in the actions logs
reporting that Node 16 is deprecated for those distros
jobs.