tv_grab_zz_sdjson: improve episode/season metadata handling (#222)
The program metadata from SD is an array of different providers. The
previous code was specifically looking for 'Gracenote' provider in the
first array element and ignoring everything else. For example, it seems the
first element is now often 'TVmaze' (at least for the lineups I use) and
the episode/season data was all being ignored.

The code now iterates over all the metadata providers preferring Gracenote
but using the first available as a fallback.

Closes #221