filter: add gap error check to tv_sort
This is mainly a development helper, i.e. to indicate to the developer
that a grabber doesn't extract all programmes and therefore causes holes
in the schedule. This is not an exact science, because there may be
legitimate gaps in the schedule.

One legitimate gap in the programme schedule occurs when a station signs
off for the day and then signs on again on the "next" day. For that
reason the --gap-error expects an integer parameter N. Any gap longer
than N minutes is considered to be OK, i.e. a gap between sign-off and
sign-on.

The --gap-error option automatically enables the --by-channel option,
because this is required for the gap error check to work.